transfered scenes: town, registerPage, loginPage, and home
